// Game-Hub/game-Hub-fe/src/app/services/models/user-settings-update-requests.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { UpdateCommunitySettingsRequest } from '../models/update-community-settings-request';
import { UpdatePrivacySettingsRequest } from '../models/update-privacy-settings-request';
import { UpdateStoreSettingsRequest } from '../models/update-store-settings-request';
export interface UserSettingsUpdateRequests {
  communitySettingsIds: Array<UpdateCommunitySettingsRequest>;
  privacySettingsIds: Array<UpdatePrivacySettingsRequest>;
  storeSettingsIds: Array<UpdateStoreSettingsRequest>;
}