// Game-Hub/game-Hub-fe/src/app/services/models/user-settings-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CardColorResponse } from '../models/card-color-response';
import { CommunitySettingsResponse } from '../models/community-settings-response';
import { CountryResponse } from '../models/country-response';
import { GenreResponse } from '../models/genre-response';
import { LevelResponse } from '../models/level-response';
import { PrivacySettingsResponse } from '../models/privacy-settings-response';
import { StoreSettingsResponse } from '../models/store-settings-response';
import { UserLibraryResponse } from '../models/user-library-response';
export interface UserSettingsResponse {
  bannerImage?: string;
  bannerType?: string;
  bio?: string;
  cardColor?: CardColorResponse;
  communitySettingsResponse?: Array<CommunitySettingsResponse>;
  country?: CountryResponse;
  email?: string;
  favoriteGame?: UserLibraryResponse;
  favoriteGenres?: Array<GenreResponse>;
  firstName?: string;
  joinedDate?: string;
  lastName?: string;
  level?: LevelResponse;
  predefinedBannerPath?: string;
  privacySettingsResponse?: Array<PrivacySettingsResponse>;
  profileColor?: string;
  storeSettingsResponse?: Array<StoreSettingsResponse>;
  userId?: number;
  userName?: string;
  userProfilePicture?: string;
}